requirements.txt
anyio==3.6.1
certifi==2022.9.24
chardet==4.0.0
click==8.1.3
fastapi==0.85.0
h11==0.14.0
idna==2.10
jsonapi-requests==0.6.2
mantiumapi==0.1.54
pydantic==1.10.2
PyJWT==2.1.0
python-dotenv==0.21.0
requests==2.25.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tenacity==7.0.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
